testing: fix import cycle in paginator test
The benchmark in the paginator test somehow includes reference cycle that hasn't
been previously detected in CI. Move this up into the `nomad/state` package.
